request的实现优化，request的h5版本复用taro-h5的实现，同时通过接口参数，业务可以灵活选择实现机制
