+++ packages/taro-platform-harmony-hybrid/src/api/apis/osChannelApi.ts
@@ -1,4 +1,4 @@
-let http
+// let http
 let display
 let wifiManager
 let call
@@ -10,7 +10,7 @@ let geoLocationManager
 
 try {
   // @ts-ignore
-  http = requireNapi('net.http')
+  // http = requireNapi('net.http')
   // @ts-ignore
   display = requireNapi('display')
   // @ts-ignore
@@ -43,181 +43,181 @@ try {
     })
 } catch (e) {}
 
-const errMsgMap = new Map([
-  [401, 'Parameter error'],
-  [201, 'Permission denied'],
-  [3, 'URL using bad/illegal format or missing URL'],
-  [7, "Couldn't connect to server"],
-  [23, 'Failed writing received data to disk/application'],
-  [25, 'Upload failed'],
-  [26, 'Failed to open/read local data from file/application'],
-  [28, 'Timeout was reached'],
-  [73, 'Remote file already exists'],
-  [78, 'Remote file not found'],
-  [999, 'Unknown Other Error'],
-])
+// const errMsgMap = new Map([
+//   [401, 'Parameter error'],
+//   [201, 'Permission denied'],
+//   [3, 'URL using bad/illegal format or missing URL'],
+//   [7, "Couldn't connect to server"],
+//   [23, 'Failed writing received data to disk/application'],
+//   [25, 'Upload failed'],
+//   [26, 'Failed to open/read local data from file/application'],
+//   [28, 'Timeout was reached'],
+//   [73, 'Remote file already exists'],
+//   [78, 'Remote file not found'],
+//   [999, 'Unknown Other Error'],
+// ])
 
 const ErrorCode = {
   PARAMETER_ERROR: 202,
 }
 
-class RequestTask {
-  private abortFlag: boolean
-  private readonly fail: (arg0: any) => any
-  private readonly complete: (arg0: any) => any
-  private readonly httpRequest: any
-  private headersCallback: Map<any, any>
-  private result: { data?: any, statusCode?: any, header?: any, cookies?: any, errMsg: string | undefined }
-  private res: { errMsg: string }
-
-  constructor (object: any) {
-    const { url, header, method = 'GET', timeout, responseType, enableCache } = object || {}
-    let { data } = object || {}
-    const { success, fail, complete } = object || {}
-    this.abortFlag = false
-    this.fail = fail
-    this.complete = complete
-    this.httpRequest = http.createHttp()
-    this.headersCallback = new Map()
-    if (!object) {
-      console.error('[OsChannel] request error: params illegal')
-      return
-    }
-    let isFormUrlEncoded = false
-    for (const key in header) {
-      if (key.toLowerCase() === 'content-type') {
-        if (header[key].toLowerCase().includes('application/x-www-form-urlencoded')) {
-          isFormUrlEncoded = true
-        }
-        break
-      }
-    }
-
-    // data为Object类型时，属性的值类型如果是number, request请求时信息会丢失. 故将data转成string类型进行规避
-    if (data && (isFormUrlEncoded || ['GET', 'OPTIONS', 'DELETE', 'TRACE', 'CONNECT'].includes(method))) {
-      const dataArray = []
-      for (const key in data) {
-        // @ts-ignore
-        dataArray.push(encodeURIComponent(key) + '=' + encodeURIComponent(data[key]))
-      }
-      data = dataArray.join('&')
-    }
-    // header的属性的值类型如果是number, request请求时信息会丢失. 故将各个属性转成string类型
-    if (header) {
-      for (const key in header) {
-        header[key] = `${header[key]}`
-      }
-    }
-    const httpRequestOptions = {
-      method: method,
-      extraData: data || {},
-      header: header,
-      connectTimeout: timeout,
-      expectDataType:
-        responseType && responseType === 'arraybuffer' ? http.HttpDataType.ARRAY_BUFFER : http.HttpDataType.STRING,
-      usingCache: enableCache || false,
-    }
-    this.httpRequest
-      .request(typeof url === 'string' ? url : '', httpRequestOptions)
-      .then((data) => {
-        if (success && !this.abortFlag) {
-          let result = data.result
-          const { dataType } = object || {}
-          if (dataType && dataType !== 'json') {
-            if (typeof data.result === 'object') {
-              result = JSON.stringify(data.result)
-            }
-          } else if (typeof data.result === 'string') {
-            try {
-              result = JSON.parse(result)
-            } catch (err) {
-              /* empty */
-            }
-          }
-          const res = {
-            data: result,
-            statusCode: data.responseCode,
-            header: data.header,
-            cookies: typeof data.cookies === 'string' ? (data.cookies ? [data.cookies] : []) : data.cookies,
-            errMsg: 'request:ok',
-          }
-          this.result = res
-          success(res)
-        }
-      })
-      .catch((err: any) => {
-        console.error('[OsChannel] request error: ' + JSON.stringify(err))
-        if (fail && !this.abortFlag) {
-          const res = {
-            errMsg: errMsgMap.has(err.code) ? errMsgMap.get(err.code) : 'Unknown Error',
-          }
-          this.result = res
-          fail(res)
-        }
-      })
-      .finally(() => {
-        if (complete && !this.abortFlag) {
-          complete(this.result)
-        }
-      })
-  }
-
-  /**
-   * interrupt request task
-   */
-  abort () {
-    this.httpRequest.destroy()
-    this.abortFlag = true
-    this.res = {
-      errMsg: 'request:fail abort',
-    }
-    this.fail && this.fail(this.res)
-    this.complete && this.complete(this.res)
-  }
-
-  /**
-   * subscribe HTTP Response Header event
-   * @callback params header {Object}: HTTP Response Header
-   */
-  onHeadersReceived (callback: any) {
-    const taskCallback = (header: any) => {
-      !this.abortFlag &&
-        callback({
-          header,
-        })
-    }
-    if (!callback) {
-      console.error('[OsChannel] Invalid, callback is null')
-      return
-    }
-    if (!this.headersCallback.has(callback)) {
-      this.headersCallback.set(callback, taskCallback)
-      if (this.httpRequest) {
-        this.httpRequest.on('headersReceive', taskCallback)
-      }
-    }
-  }
-
-  /**
-   * unsubscribe HTTP Response Header event
-   * remove all if callback is null, otherwise remove the specialized callback
-   */
-  offHeadersReceived (callback: any) {
-    if (!callback) {
-      this.headersCallback.clear()
-      if (this.httpRequest) {
-        this.httpRequest.off('headersReceive')
-      }
-    } else if (this.headersCallback.has(callback)) {
-      if (this.httpRequest) {
-        this.httpRequest.off('headersReceive', this.headersCallback.get(callback))
-      }
-      this.headersCallback.delete(callback)
-    } else {
-      /* empty */
-    }
-  }
-}
+// class RequestTask {
+//   private abortFlag: boolean
+//   private readonly fail: (arg0: any) => any
+//   private readonly complete: (arg0: any) => any
+//   private readonly httpRequest: any
+//   private headersCallback: Map<any, any>
+//   private result: { data?: any, statusCode?: any, header?: any, cookies?: any, errMsg: string | undefined }
+//   private res: { errMsg: string }
+//
+//   constructor (object: any) {
+//     const { url, header, method = 'GET', timeout, responseType, enableCache } = object || {}
+//     let { data } = object || {}
+//     const { success, fail, complete } = object || {}
+//     this.abortFlag = false
+//     this.fail = fail
+//     this.complete = complete
+//     this.httpRequest = http.createHttp()
+//     this.headersCallback = new Map()
+//     if (!object) {
+//       console.error('[OsChannel] request error: params illegal')
+//       return
+//     }
+//     let isFormUrlEncoded = false
+//     for (const key in header) {
+//       if (key.toLowerCase() === 'content-type') {
+//         if (header[key].toLowerCase().includes('application/x-www-form-urlencoded')) {
+//           isFormUrlEncoded = true
+//         }
+//         break
+//       }
+//     }
+//
+//     // data为Object类型时，属性的值类型如果是number, request请求时信息会丢失. 故将data转成string类型进行规避
+//     if (data && (isFormUrlEncoded || ['GET', 'OPTIONS', 'DELETE', 'TRACE', 'CONNECT'].includes(method))) {
+//       const dataArray = []
+//       for (const key in data) {
+//         // @ts-ignore
+//         dataArray.push(encodeURIComponent(key) + '=' + encodeURIComponent(data[key]))
+//       }
+//       data = dataArray.join('&')
+//     }
+//     // header的属性的值类型如果是number, request请求时信息会丢失. 故将各个属性转成string类型
+//     if (header) {
+//       for (const key in header) {
+//         header[key] = `${header[key]}`
+//       }
+//     }
+//     const httpRequestOptions = {
+//       method: method,
+//       extraData: data || {},
+//       header: header,
+//       connectTimeout: timeout,
+//       expectDataType:
+//         responseType && responseType === 'arraybuffer' ? http.HttpDataType.ARRAY_BUFFER : http.HttpDataType.STRING,
+//       usingCache: enableCache || false,
+//     }
+//     this.httpRequest
+//       .request(typeof url === 'string' ? url : '', httpRequestOptions)
+//       .then((data) => {
+//         if (success && !this.abortFlag) {
+//           let result = data.result
+//           const { dataType } = object || {}
+//           if (dataType && dataType !== 'json') {
+//             if (typeof data.result === 'object') {
+//               result = JSON.stringify(data.result)
+//             }
+//           } else if (typeof data.result === 'string') {
+//             try {
+//               result = JSON.parse(result)
+//             } catch (err) {
+//               /* empty */
+//             }
+//           }
+//           const res = {
+//             data: result,
+//             statusCode: data.responseCode,
+//             header: data.header,
+//             cookies: typeof data.cookies === 'string' ? (data.cookies ? [data.cookies] : []) : data.cookies,
+//             errMsg: 'request:ok',
+//           }
+//           this.result = res
+//           success(res)
+//         }
+//       })
+//       .catch((err: any) => {
+//         console.error('[OsChannel] request error: ' + JSON.stringify(err))
+//         if (fail && !this.abortFlag) {
+//           const res = {
+//             errMsg: errMsgMap.has(err.code) ? errMsgMap.get(err.code) : 'Unknown Error',
+//           }
+//           this.result = res
+//           fail(res)
+//         }
+//       })
+//       .finally(() => {
+//         if (complete && !this.abortFlag) {
+//           complete(this.result)
+//         }
+//       })
+//   }
+//
+//   /**
+//    * interrupt request task
+//    */
+//   abort () {
+//     this.httpRequest.destroy()
+//     this.abortFlag = true
+//     this.res = {
+//       errMsg: 'request:fail abort',
+//     }
+//     this.fail && this.fail(this.res)
+//     this.complete && this.complete(this.res)
+//   }
+//
+//   /**
+//    * subscribe HTTP Response Header event
+//    * @callback params header {Object}: HTTP Response Header
+//    */
+//   onHeadersReceived (callback: any) {
+//     const taskCallback = (header: any) => {
+//       !this.abortFlag &&
+//         callback({
+//           header,
+//         })
+//     }
+//     if (!callback) {
+//       console.error('[OsChannel] Invalid, callback is null')
+//       return
+//     }
+//     if (!this.headersCallback.has(callback)) {
+//       this.headersCallback.set(callback, taskCallback)
+//       if (this.httpRequest) {
+//         this.httpRequest.on('headersReceive', taskCallback)
+//       }
+//     }
+//   }
+//
+//   /**
+//    * unsubscribe HTTP Response Header event
+//    * remove all if callback is null, otherwise remove the specialized callback
+//    */
+//   offHeadersReceived (callback: any) {
+//     if (!callback) {
+//       this.headersCallback.clear()
+//       if (this.httpRequest) {
+//         this.httpRequest.off('headersReceive')
+//       }
+//     } else if (this.headersCallback.has(callback)) {
+//       if (this.httpRequest) {
+//         this.httpRequest.off('headersReceive', this.headersCallback.get(callback))
+//       }
+//       this.headersCallback.delete(callback)
+//     } else {
+//       /* empty */
+//     }
+//   }
+// }
 
 const hideKeyboard = () => {
   return new Promise((resolve, reject) => {
@@ -372,26 +372,26 @@ const getAppAuthorizeSetting = () => {
   return result
 }
 
-const request = (options: any) => {
-  const { success, fail, complete } = options
-  const task = new RequestTask(options)
-  // if success / fail / complete, return requestTask otherwise return Promise<requestTask>
-  if (options && !success && !fail && !complete) {
-    success(task)
-    return
-  }
-  if (!options) {
-    fail(['illegal params', -1])
-  }
-  return task
-}
+// const request = (options: any) => {
+//   const { success, fail, complete } = options
+//   const task = new RequestTask(options)
+//   // if success / fail / complete, return requestTask otherwise return Promise<requestTask>
+//   if (options && !success && !fail && !complete) {
+//     success(task)
+//     return
+//   }
+//   if (!options) {
+//     fail(['illegal params', -1])
+//   }
+//   return task
+// }
 
 const osChannelApi = {
   getSystemSetting,
   hideKeyboard,
   makePhoneCall,
   getAppAuthorizeSetting,
-  request,
+  // request,
 }
 
 export default osChannelApi
